# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Car Rental API",
        "description": "Dokumentacja systemu wypożyczalni",
        "version": "1.0.0"
    },
    "paths": {
...
        "/admin/rentals": {
            "get": {
                "tags": [
                    "Admin Rentals"
                ],
                "summary": "Lista wszystkich wynajmów",
                "description": "Zwraca paginowaną listę wynajmów z możliwością filtrowania po statusie, użytkowniku lub punkcie.",
                "operationId": "getAdminRentals",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "user_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "rental_point_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginowana lista wynajmów"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin Rentals"
                ],
                "summary": "Utwórz nowy wynajem/rezerwację (ręcznie)",
                "operationId": "storeAdminRental",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "car_id",
                                    "rental_point_start_id",
                                    "rental_point_end_id",
                                    "start_date",
                                    "planned_end_date"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "car_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "rental_point_start_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "rental_point_end_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "start_date": {
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "planned_end_date": {
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "use_extra_insurance": {
                                        "type": "boolean",
                                        "default": false
                                    },
                                    "notes": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Wynajem utworzony"
                    },
                    "422": {
                        "description": "Błąd walidacji, brak środków lub auto zajęte"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/rentals/{id}": {
            "get": {
                "tags": [
                    "Admin Rentals"
                ],
                "summary": "Szczegóły konkretnego wynajmu",
                "operationId": "getAdminRentalById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pełne dane o wynajmie i transakcjach"
                    },
                    "404": {
                        "description": "Nie znaleziono wynajmu"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Admin Rentals"
                ],
                "summary": "Usuń rekord wynajmu",
                "operationId": "deleteAdminRental",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Usunięto"
                    },
                    "422": {
                        "description": "Nie można usunąć aktywnego wynajmu"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Admin Rentals"
                ],
                "summary": "Edycja danych wynajmu",
                "operationId": "updateAdminRental",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "active",
                                            "completed",
                                            "cancelled",
                                            "early_return"
                                        ]
                                    },
                                    "notes": {
                                        "type": "string"
                                    },
                                    "rental_point_end_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Zaktualizowano"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/rentals/{id}/cancel": {
            "post": {
                "tags": [
                    "Admin Rentals"
                ],
                "summary": "Anulowanie rezerwacji lub wczesny zwrot",
                "operationId": "cancelAdminRental",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "cancellation_reason"
                                ],
                                "properties": {
                                    "cancellation_reason": {
                                        "type": "string",
                                        "example": "Zmiana planów"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Sukces - zwrócono środki"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/rentals/{id}/approve-return": {
            "post": {
                "tags": [
                    "Admin Rentals"
                ],
                "summary": "Zatwierdzenie zwrotu auta",
                "operationId": "approveAdminRentalReturn",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "notes": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Zwrot zatwierdzony"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/rentals/{id}/reject-return": {
            "post": {
                "tags": [
                    "Admin Rentals"
                ],
                "summary": "Odrzucenie zwrotu auta",
                "operationId": "rejectAdminRentalReturn",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "reason": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Zwrot odrzucony"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
...
}